requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Pillow==11.0.0